Fix Dockerfile for new Python base image
For some reason, the new Python base image refuses to install the
requirements properly.

Made a workaround for this by running pip ourselves, and making sure to:
a) upgrade pbr to the latest version
b) do not use --no-cache-dir with pip

No idea why this works, since pbr is eventually downgraded when
installing the application requirements. Maybe a bug in pip / pbr.